:recycle: tidy up resolver
