+++ graph/resolver.go
@@ -1,6 +1,13 @@
 package graph
 
-import "github.com/do-it-if-i-can/server/graph/model"
+import (
+	"context"
+	"fmt"
+	"math/rand"
+
+	"github.com/do-it-if-i-can/server/graph/generated"
+	"github.com/do-it-if-i-can/server/graph/model"
+)
 
 // This file will not be regenerated automatically.
 //
@@ -8,4 +15,28 @@ import "github.com/do-it-if-i-can/server/graph/model"
 
 type Resolver struct {
 	todos []*model.Todo
+}
+
+func (r *Resolver) Mutation() generated.MutationResolver { return &mutationResolver{r} }
+func (r *Resolver) Query() generated.QueryResolver       { return &queryResolver{r} }
+
+type mutationResolver struct{ *Resolver }
+type queryResolver struct{ *Resolver }
+
+// mutation ----------------------------------------------
+
+func (r *mutationResolver) CreateTodo(ctx context.Context, input model.NewTodo) (*model.Todo, error) {
+	todo := &model.Todo{
+		Title:  input.Title,
+		ID:     fmt.Sprintf("T%d", rand.Int()),
+		UserID: input.UserID, // fix this line
+	}
+	r.todos = append(r.todos, todo)
+	return todo, nil
+}
+
+// query  ----------------------------------------------
+
+func (r *queryResolver) Todos(ctx context.Context, category *model.Category) ([]*model.Todo, error) {
+	return r.todos, nil
 }
